requirements.txt
asgiref==3.6.0
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.4
decorator==5.1.1
defusedxml==0.7.1
Django==3.2.16
django-filter==22.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
executing==1.2.0
idna==3.4
inflection==0.5.1
iniconfig==1.1.1
ipython==8.11.0
itypes==1.2.0
jedi==0.18.2
Jinja2==3.1.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
oauthlib==3.2.2
packaging==22.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pluggy==0.13.1
prompt-toolkit==3.0.38
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pycparser==2.21
Pygments==2.14.0
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python3-openid==3.2.0
pytz==2022.7
requests==2.26.0
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
stack-data==0.6.2
toml==0.10.2
traitlets==5.9.0
uritemplate==4.1.1
urllib3==1.26.13
wcwidth==0.2.6
